+++ mutt_socket.c
@@ -31,28 +31,11 @@
 #include "url.h"
 
 /* support for multiple socket connections */
-static struct ConnectionList Connections = TAILQ_HEAD_INITIALIZER(Connections);
+struct ConnectionList Connections = TAILQ_HEAD_INITIALIZER(Connections);
 
 struct ConnectionList *mutt_socket_head(void)
 {
   return &Connections;
-}
-
-/**
- * mutt_socket_free - remove connection from connection list and free it
- */
-void mutt_socket_free(struct Connection *conn)
-{
-  struct Connection *np = NULL;
-  TAILQ_FOREACH(np, &Connections, entries)
-  {
-    if (np == conn)
-    {
-      TAILQ_REMOVE(&Connections, np, entries);
-      FREE(&np);
-      return;
-    }
-  }
 }
 
 /**
